+++ ihp/cells/__init__.py
@@ -13,3 +13,4 @@ from .bjt_transistors import *
 from .inductors import *
 from .resistors import *
 from .antennas import *
+from .rf_devices import *
